+++ server/index.ts
@@ -1,5 +1,5 @@
 import { Meteor } from 'meteor/meteor'
-import { check } from 'meteor/check'
+import { check, Match } from 'meteor/check'
 import {
   APP_NEWS,
   FlashNewsCollection,
@@ -23,31 +23,38 @@ export {
 /**
  * Gets current flash news for the site
  * @param limit {Number}
+ * @param language {String}
  * @return {Mongo.Cursor}
  */
-Meteor.publish('freedombase:flashnews-getMain', (limit = 3) => {
-  check(limit, Number)
-  return FlashNewsCollection.find(
-    {
-      ...currentFlashNewsSelector,
-      objectType: APP_NEWS
-    },
-    { limit }
-  )
-})
+Meteor.publish(
+  'freedombase:flashnews-getMain',
+  (limit = 3, language = 'en') => {
+    check(limit, Match.Optional(Number))
+    check(language, Match.Optional(String))
+    return FlashNewsCollection.find(
+      {
+        ...currentFlashNewsSelector,
+        objectType: APP_NEWS
+      },
+      { limit }
+    )
+  }
+)
 /**
  * Gets current flash news for the given object
  * @param objectType {String}
  * @param objectId {String}
  * @param limit {Number}
+ * @param language {String}
  * @returns {Mongo.Cursor}
  */
 Meteor.publish(
   'freedombase:flashnews-getFor',
-  (objectType: String, objectId: String, limit = 5) => {
+  function (objectType: String, objectId: String, limit = 5, language = 'en') {
     check(objectType, String)
     check(objectId, String)
-    check(limit, Number)
+    check(limit, Match.Optional(Number))
+    check(language, Match.Optional(String))
     return FlashNewsCollection.find({
       ...currentFlashNewsSelector,
       objectType,