+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 # Changelog
 
 ## Next - Unreleased
+
 ### Breaking changes
+
 ### New features
+
 ### Fixes
+
 ### Depracated
 
+## v0.2.0 - 2021-11-23
+
+### Breaking changes & New features
+
+* Added option to make news localized
+* Now includes `socialize:base-model`
+
 ## v0.1.0 - 2021-11-19
+
 * Initial testing release